Move investment controllers under sub-app
