+++ src/sass/components/_local-nav.scss
@@ -35,3 +35,8 @@
     font-weight: 600;
   }
 }
+
+.local-nav--background .is-active {
+  background-color: $grey-1;
+}
+
