+++ src/server.js
@@ -26,9 +26,6 @@ const apiController = require('./controllers/api.controller')
 const contactController = require('./controllers/contact.controller')
 const contactEditController = require('./controllers/contact-edit.controller')
 const contactInteractionController = require('./controllers/contact-interaction.controller')
-const investmentStartController = require('./controllers/investment-start.controller')
-const investmentCreateController = require('./controllers/investment-create.controller')
-const investmentDetailsController = require('./controllers/investment-details.controller')
 const indexController = require('./controllers/index.controller')
 const loginController = require('./controllers/login.controller')
 const myAccountController = require('./controllers/my-account.controller')
@@ -133,9 +130,7 @@ app.use(require('./controllers/company-contact.controller').router)
 app.use(require('./controllers/company-add.controller').router)
 app.use(require('./controllers/company-archive.controller').router)
 app.use(require('./controllers/company-export.controller').router)
-app.use(investmentStartController.router)
-app.use(investmentCreateController.router)
-app.use(investmentDetailsController.router)
+app.use(require('./controllers/investment').router)
 app.use(contactController.router)
 app.use(require('./controllers/contact-archive.controller').router)
 app.use(contactEditController.router)
