+++ src/views/investment/_form.njk
@@ -1,0 +1,194 @@
+<form method="post">
+  <input type="hidden" name="_csrf" value="{{ csrfToken }}">
+  <input type="hidden" name="investor_company" value="{{ equityCompany.id }}">
+
+  {{ trade.autocomplete('client_contacts',
+    label="Client contact for this project",
+    value=form.state['client_contacts'],
+    options=form.options.contacts)
+  }}
+
+  {% component 'form-multiple-choice', {
+    variant: 'radio',
+    label: 'Are you the client relationship manager for this project?',
+    name: 'is-relationship-manager',
+    inline: true,
+    value: form.state['is-relationship-manager'],
+    children: [{
+      label: 'Yes',
+      value: user.id
+    }, {
+      label: 'No'
+    }]
+  } %}
+
+  <div class="panel panel-border-narrow js-ConditionalSubfield" data-controlled-by="is-relationship-manager" data-control-value="No">
+    {{ trade.autocomplete('client_relationship_manager',
+      label="Client relationship manager",
+      value=form.state['client_relationship_manager'],
+      options=form.options.advisers)
+    }}
+  </div>
+
+  {% component 'form-multiple-choice', {
+    variant: 'radio',
+    label: 'Are you the referral source for this project?',
+    name: 'is-referral-source',
+    inline: true,
+    value: form.state['is-referral-source'],
+    children: [{
+      label: 'Yes',
+      value: user.id
+    }, {
+      label: 'No'
+    }]
+  } %}
+
+  <div class="panel panel-border-narrow js-ConditionalSubfield" data-controlled-by="is-referral-source" data-control-value="No">
+    {{ trade.autocomplete('referral_source_adviser',
+      label="Referral source adviser",
+      value=form.state['referral_source_adviser'],
+      options=form.options.advisers)
+    }}
+  </div>
+
+  {{ trade.dropdown('referral_source_activity',
+    label='Referral source activity',
+    value=form.state['referral_source_activity'],
+    emptyLabel='Choose referral source activity',
+    options=form.options.referralSourceActivities)
+  }}
+
+  <div class="panel panel-border-narrow js-ConditionalSubfield" data-controlled-by="referral_source_activity" data-control-value="{{ form.options.referralSourceActivities[2].id }}">
+    {{ trade.textbox('referral_source_activity_event',
+      value=form.state['referral_source_activity_event'],
+      label='Event')
+    }}
+  </div>
+
+  <div class="panel panel-border-narrow js-ConditionalSubfield" data-controlled-by="referral_source_activity" data-control-value="{{ form.options.referralSourceActivities[3].id }}">
+    {{ trade.dropdown('referral_source_activity_marketing',
+      label='Marketing',
+      value=form.state['referral_source_activity_marketing'],
+      emptyLabel='Choose a marketing',
+      options=form.options.referralSourceMarketing)
+    }}
+  </div>
+
+  <div class="panel panel-border-narrow js-ConditionalSubfield" data-controlled-by="referral_source_activity" data-control-value="{{ form.options.referralSourceActivities[9].id }}">
+    {{ trade.dropdown('referral_source_activity_website',
+      label='Website',
+      value=form.state['referral_source_activity_website'],
+      emptyLabel='Choose a website',
+      options=form.options.referralSourceWebsite)
+    }}
+  </div>
+
+  {% component 'form-multiple-choice', {
+    variant: 'radio',
+    label: 'Type of investment',
+    name: 'investment_type',
+    value: form.state['investment_type'],
+    children: form.options.investmentTypes
+  } %}
+
+  <div class="panel panel-border-narrow js-ConditionalSubfield" data-controlled-by="investment_type" data-control-value="{{ form.options.investmentTypes[1].value }}">
+    {{ trade.dropdown('fdi_type',
+      label='Type of foreign direct investment (FDI)',
+      value=form.state['fdi_type'],
+      emptyLabel='Choose a type',
+      options=form.options.fdi)
+    }}
+  </div>
+
+  <div class="panel panel-border-narrow js-ConditionalSubfield" data-controlled-by="investment_type" data-control-value="{{ form.options.investmentTypes[2].value }}">
+    {{ trade.dropdown('non_fdi_type',
+      label='Type of non-foreign direct investment (Non-FDI)',
+      value=form.state['non_fdi_type'],
+      emptyLabel='Choose a type',
+      options=form.options.nonFdi)
+    }}
+  </div>
+
+  {{ trade.dropdown('sector',
+    label='Primary sector',
+    value=form.state['sector'],
+    emptyLabel='Choose a sector',
+    options=form.options.primarySectors)
+  }}
+
+  {{ trade.dropdown('business_activities',
+    label='Business activity of this investment project',
+    value=form.state['business_activities'],
+    emptyLabel='Choose a business activity',
+    options=form.options.businessActivities)
+  }}
+
+  {{ trade.textbox('name',
+    value=form.state['name'],
+    label='Project name')
+  }}
+
+  {{ trade.textarea('description',
+    value=form.state['description'],
+    label='Description')
+  }}
+
+  {% component 'form-multiple-choice', {
+    variant: 'radio',
+    label: 'Can this project be shared with our local partner network, outside of DIT?',
+    name: 'project_shareable',
+    inline: true,
+    value: form.state['project_shareable'],
+    children: [{
+      label: 'Yes',
+      value: 'true'
+    }, {
+      label: 'No',
+      value: 'false'
+    }]
+  } %}
+
+  {% component 'form-multiple-choice', {
+    variant: 'radio',
+    label: 'Have we signed a non-disclosure agreement (NDA) with the client?',
+    name: 'nda_signed',
+    inline: true,
+    value: form.state['nda_signed'],
+    children: [{
+      label: 'Yes',
+      value: 'true'
+    }, {
+      label: 'No',
+      value: 'false'
+    }]
+  } %}
+
+  <div id="land-date-wrapper" class="form-group form-date ">
+    <fieldset>
+      <legend class="form-label-bold">Estimated land date</legend>
+      <span class="form-hint">When activities planned under the investment project will have fully commenced</span>
+
+      <div class="form-group form-group-month">
+        <label for="land-date_month">
+          <span class="form-hint">MM</span>
+        </label>
+        <input type="text" id="land-date_month" class="form-control" name="land-date_month" value="{{ form.state['land-date_month'] }}" maxlength="2" autocomplete="mm">
+      </div>
+
+      <div class="form-group form-group-year">
+        <label for="land-date_year">
+          <span class="form-hint">YYYY</span>
+        </label>
+        <input type="text" id="land-date_year" class="form-control" name="land-date_year" value="{{ form.state['land-date_year'] }}" maxlength="4" autocomplete="yyyy">
+      </div>
+    </fieldset>
+  </div>
+
+  {% if investmentProject %}
+    {% set backUrl = '/investment/' + investmentProject.meta.id %}
+  {% else %}
+    {% set backUrl = '/investment/start?client-company=' + equityCompany.id %}
+  {% endif %}
+  {{ trade.save(backUrl=backUrl) }}
+</form>
