+++ src/views/investment/_layout.njk
@@ -1,8 +1,28 @@
 {% extends "layouts/_base-two-column.njk" %}
 
-{% block main_heading %}{% endblock %}
+{% block main_heading %}
+  {% if investmentProject.meta %}
+    <h1 class="visually-hidden">Investment details</h1>
+    <h2 class="page-heading">{{ investmentProject.meta.name }}</h2>
+    {% component 'project-status', investmentProject.meta %}
+  {% endif %}
+{% endblock %}
 
-{% block main_grid_left_column %}{% endblock %}
+{% block main_grid_left_column %}
+  {% if localNavItems %}
+    <nav class="local-nav {{ 'local-nav--background' if variant == 'edit' }}">
+      {% for item in localNavItems %}
+        {% set isSelected = currentNavItem == item.slug %}
+        <a
+          href="/investment/{{ investmentProject.meta.id }}/{{ item.slug }}"
+          class="local-nav__link {{ 'is-active' if isSelected}}"
+        >
+          {{ item.label }}
+        </a>
+      {% endfor %}
+    </nav>
+  {% endif %}
+{% endblock %}
 
 {% block main_grid_right_column %}{% endblock %}
