+++ src/views/investment/start.njk
@@ -5,6 +5,8 @@
   <h1 class="page-heading">Add new investment project</h1>
 {% endblock %}
 
+{% block main_grid_left_column %}{% endblock %}
+
 {% block main_grid_right_column %}
   {% if clientCompany %}
     <h2 class="heading-medium">Client company</h2>
@@ -114,7 +116,7 @@
       <p class="infostrip">If you can’t find the company you’re looking for, <a href="/company/add-step-1">add a new compay</a>. Unfortunately you will lose your progress.</p>
     {% endif %}
 
-  {% else %}
+  {% elseif clientCompany %}
 
     <form class="section archive-panel__form" method="post">
       <input type="hidden" name="_csrf" value="{{ csrfToken }}">
@@ -141,5 +143,12 @@
       {{ trade.save(buttonText="Continue") }}
     </form>
 
+  {% else %}
+
+    <p class="flash flash--info">
+      Client company is required
+    </p>
+
   {% endif %}
+
 {% endblock %}
